+++ android/src/main/java/com/sean/rao/ali_auth/common/CustomAuthUIControlClickListener.java
@@ -44,12 +44,10 @@ public class CustomAuthUIControlClickListener extends LoginParams implements Aut
       //点击授权页默认样式的返回按钮
       case ResultCode.CODE_ERROR_USER_CANCEL:
         Log.e(TAG, "点击了授权页默认返回按钮");
-        mAuthHelper.quitLoginPage();
         break;
       //点击授权页默认样式的切换其他登录方式 会关闭授权页
       //如果不希望关闭授权页那就setSwitchAccHidden(true)隐藏默认的  通过自定义view添加自己的
       case ResultCode.CODE_ERROR_USER_SWITCH:
-        mAuthHelper.quitLoginPage();
         Log.e(TAG, "点击了授权页默认切换其他登录方式");
         break;
       //点击一键登录按钮会发出此回调