+++ example/lib/my_home_page.dart
@@ -53,7 +53,7 @@ class MyHomePageState extends State<MyHomePage> with WidgetsBindingObserver {
     /// 初始化第三方按钮数据
     setState(() {
       androidSk =
-          "kReOFbE7mWUhefnT2BAGdKCGqgtnZbcPa/vGFc6z//ytIrOxZRndH0pdIeHM465OfdpHPMFxgPATR1EnKE7aonZ5hyEP1CE4Wz0QhWEWNTg7mmeq8hstbgFtjP8boZx/mPalQZmfD5heQ9E5Rahg4tWQfsCBENlQLgR/6vqtA8F3knXFa6awGegHj3C8bSXyCVj2OKxZFAvrZ1+1bd7TD2We3HyXsSJoDBLGuSqZIZ3VkExNC8jX4fL9uP5Ul9VVNSjahTn70u+9RYdB0BrtJpw+FPytOIsapzqdfrtqkFTk3v0+BcVJnA==";
+          "RwZ0kq2Zn+0IQFb0/1ajfhpYg9qmw/VyRLuumsC5TtVHkyncsQkxjghtBJ70ZxsuiJD/LF7Sp+Be/i/QBGfdSq1unyvuZz/4Zq1B61beqnK9AkEWGM7fJ2N+yNWtBUQ8vAOJqu6+WA+R0+/eG3c5lyUaudycPMaw7fNm7wjR2xCPYCzkYrPC0sA1HWW+s2Rp32DcP87IjftvLnRIuCa/OGk2yPWaENqW68MF0EF2S8M0gscaamkQ+pHR/+R/4jFNYt1jPKEMx9Jqxuy96uRC0WoOPaFJ428KdRQbG7AgLFENEgFb9wkq5A==";
       iosSk =
           "mjWr9sTsoXwmMx7qf0T2KQOQBpqkxeNW9I1ZNZ96ZCeBbeD9xYOUaC2mE9mcqog041VCot2sLcy9UArf+re517e5R9yowKCjf15VglZSP/HweRhOT8Cvci43zagyRqo40l85LTnZ5uJPaVauDLJB7hOTIkNPGm3fb621k6A6ZDh6aDGAKWyy0tPUPV/9RFrfeig9SURNe9Vl/Aok6SKg+SftM30uk2W8wdbV8gMVbU51Odnoapm2ZlAJYmCrdoXvROW5qc8pbQ8=";
 
@@ -103,6 +103,15 @@ class MyHomePageState extends State<MyHomePage> with WidgetsBindingObserver {
         print("----------------> $onEvent <----------------");
       }
 
+      // 自己关闭授权页面
+      if (onEvent["code"] == "700005") {
+        AliAuth.quitPage();
+      }
+
+      // 自己关闭授权页面
+      if (onEvent["code"] == "600000" && onEvent["data"] != null) {
+        // AliAuth.quitPage();
+      }
       // EasyLoading.show(status: onEvent['msg'] ?? "", maskType: EasyLoadingMaskType.black);
       setState(() {
         status = onEvent.toString();
@@ -316,7 +325,7 @@ class MyHomePageState extends State<MyHomePage> with WidgetsBindingObserver {
       androidSk,
       iosSk,
       isDebug: true,
-      autoQuitPage: true,
+      autoQuitPage: false,
       isDelay: isDelay,
       pageType: PageType.fullPort,
       statusBarColor: "#026ED2",
@@ -412,7 +421,7 @@ class MyHomePageState extends State<MyHomePage> with WidgetsBindingObserver {
       authPageActOut: "in_activity",
       activityIn: "out_activity",
       screenOrientation: -1,
-      logBtnToastHidden: false,
+      logBtnToastHidden: true,
       dialogAlpha: 1.0,
       privacyOperatorIndex: 0,
       /**